requirements.txt
asgiref==3.5.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Django==4.0.5
django-jazzmin==2.5.0
django-jet==1.0.8
djangorestframework==3.13.1
ecdsa==0.18.0
fido2==1.1.1
gunicorn==20.1.0
idna==3.4
intelhex==2.3.0
Pillow==9.2.0
pycparser==2.21
pyserial==3.5
pytz==2022.2.1
pyusb==1.2.1
requests==2.31.0
six==1.16.0
solo-python==0.1.1
solo1==0.1.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==2.0.3
whitenoise==6.2.0
